// web/saas-admin/src/lib/types.ts
export interface PackagePlan {
  code: string
  description: string
  limits: Record<string, number>
  name: string
  status: number
  version: number
}

export interface PackagesData {
  packages: PackagePlan[]
}